refactor(aptl-mcp-common): split ssh.ts below the 500-line S104 budget
Split the 1168-line ssh.ts into a shared contracts leaf (ssh-contracts.ts),
the PersistentSession lifecycle module (ssh-session.ts), and the
SSHConnectionManager orchestration module (ssh-manager.ts), with ssh.ts
reduced to a thin compatibility facade re-exporting the exact public surface.
Public exports and runtime behavior are unchanged; every source file is now
under the 500 non-comment-line S104 budget.
