+++ changelog.d/790.changed.md
@@ -1,0 +1,5 @@
+`ssh.ts` in `aptl-mcp-common` is split into a shared contracts leaf, the
+`PersistentSession` lifecycle module, and the `SSHConnectionManager`
+orchestration module, behind a thin `ssh.ts` compatibility facade. Public
+exports and runtime behavior are unchanged; the split brings every source
+file below the 500-line budget.
